requirements.txt
numpy
tensorflow
keras
matplotlib
opencv-python
scikit-learn
pandas
scikit-image
pillow
requests
flask
google-cloud-dialogflow
python-dotenv
pusher